[secure] move safety rules net to a general purpose net
This will be a temporary stand-in while working with the networking team
to figure out a better long term solution.
This is just moving it to a general purpose location for reuse across
other services.
