+++ consensus/safety-rules/src/tests/mod.rs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 mod local;
-mod network;
 mod safety_rules;
 mod serializer;
 mod spawned_process;
