+++ secure/net/src/lib.rs
@@ -1,7 +1,7 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-#![allow(dead_code)]
+#![forbid(unsafe_code)]
 
 //! This provides a simple networking substrate between a client and server. It is assumed that all
 //! operations are blocking and return only complete blocks of data. The intended use case has the
@@ -178,5 +178,30 @@ impl NetworkStream {
             unwritten = &data[total_written..];
         }
         Ok(())
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use libra_config::utils;
+    use std::net::{IpAddr, Ipv4Addr, SocketAddr};
+
+    #[test]
+    fn test_ping() {
+        let server_port = utils::get_available_port();
+        let server_addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), server_port);
+        let mut server = NetworkServer::new(server_addr);
+        let mut client = NetworkClient::connect(server_addr).unwrap();
+
+        let data = vec![0, 1, 2, 3];
+        client.write(&data).unwrap();
+        let result = server.read().unwrap();
+        assert_eq!(data, result);
+
+        let data = vec![4, 5, 6, 7];
+        server.write(&data).unwrap();
+        let result = client.read().unwrap();
+        assert_eq!(data, result);
     }
 }
